Modularizes controllers
