+++ lib/devise/controllers/base.rb
@@ -1,0 +1,12 @@
+module Devise
+  module Controllers
+    module Base
+      def self.included(base)
+        base.class_eval do
+
+        end
+      end
+
+    end
+  end
+end
